Added cases edition functionality to lawyer's view

+++ src/main/java/com/example/repository/CaseRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 
 @Repository("caseRepository")
 public interface CaseRepository extends JpaRepository<Case,Long> {
-   Case findById(Integer Id);
+   Case findById(Long Id);
 }
